Fix TinyMCE relative links
(e.g. "/replies/23828" which would be "23828" if you're on a /replies/
page and using the editor, or "../replies/23828" if you're on a
/posts/ page and using it.)
Should now truncate the host if it's on the specified URL and do an
absolute path. Would need to be altered if transferred to another
domain.
See here for reference:
https://www.tinymce.com/docs/get-started/get-support/#absoluteurls
(path2/file.htm >> /path1/path2/file.htm)