Make the <link rel="canonical"> thing show when logged in too
I could've sworn I had a good reason for it but I seem to have totally
forgotten why I had it only on log-out. I can't seem to come up with
the reason again and I'm currently thinking it's best just to have
the tag available since it's a good standard to have?