Display 'No posts yet' if no visible posts, not just if no posts
Check using posts.any? for visible_to? current_user
This should not load extra posts as .any? should limit itself
to the paginated posts. The list itself checks there aren't
plural pages of posts to look at, else .any? could return
that there aren't any posts, whereas the pagination should
in fact be displayed because there are (somehow) 25 posts
that are not visible to the current user, but could be other
pages.
The individual pages that check for any visible posts do not
do this since I decided it was an unlikely scenario. :P