Fix reply#preview then using replies_path
And associated fixes.

Adelene reported an issue in which editing a reply, pressing preview,
and then saving would create a new reply. I confirmed that it happened
(every time), investigated the code, and didn't really understand why
it was being done the way it was. This makes it use the inbuilt system
to handle the paths instead of manually setting it (by retaining the ID
on the replies/posts).

I tested it manually and everything seems to work (Reply/Post saving
from #new, saving from #preview, previewing from #preview and then
saving from there).