Fix the arrow on Tags Owed with fully unread posts
Previously, if a post was fully unread, it wouldn't count
in the `unread_post?` method (as that's designed to know
whether to display a link arrow). As a result, the arrow
would display as though you had fully read a post if you
had marked it completely unread. This checks through
`opened_ids` (via `opened_post?`) if the `@show_unread`
parameter is true, so it will show the arrow green on
fully unread posts too.