Set cookies_serializer to allow transparent upgrade to JSON
This file is explicitly mentioned in the 4.0 -> 4.1 upgrade
and it says if you want to use the new JSON format to set
:hybrid here, to transparently migrate cookies; I'm not sure
if this file was just ignored before, and we weren't using
JSON, or what?

If I run the server without this change it crashes when I try
loading the site, and with this change it works.