Fix the wide info box resulting from long content names
Long content names in `left-info-box`es (e.g. long character names,
screennames without spaces) can cause the table to expand and
accordingly break the display with the content on the right.

This fixes that by forcing the left table width to be fixed, instead
of adjusting to ignore the width given in the CSS.