Modify Presentable to use Module#prepend
This commit modifies the behavior of Presentable, but in a way that does
not affect extant code.

Including Presentable can now be done at any point in the model, and
will not change behavior based on this. It effectively causes the model
to inherit from the new Presents module, which implements `as_json` and
`as_json_without_presenter`.

This means the `as_json` methods in models can now use `super` to refer
to the Presentable method (or super(options.merge(without_presenter:
true)) to refer to the super non-Presentable method), which should make
writing code around this more understandable.