Hack that should eventually be worked around
Setting the join `*_tag` lines that refer to `:tag` to have `optional:
false` fails.

I am not sure why Rails knows the post has content_warnings, and it has
post_tags to match, but the post_tag for the new content warning has no
tag attached and so it errors. It seems like it should know it's the new
one. I cannot find a relevant place to add `inverse_of` to fix it.