Fix Dreamwidth scraping if a broken-depth comment is encountered
This only fixes it for when the broken-depth comment isn't on the page
boundary, which is 24/25 of the time. A fix is going to be done for the
other example.

See here:
https://alicornutopia.dreamwidth.org/22671.html?thread=14698127&style=site#cmt14698127

The second and third comments are on the same depth, and accordingly
break the old system's way of figuring out the pages of the thread (by
depth). Instead, use every 25th comment as a page boundary.