Override layout-th styling in post-content tables
We use `th` semantically all over the place for headers, and then style
them a particular way for consistent formatting. Unfortunately, this
messes with `th` being used in replies, since it messes up the
formatting. Unset the changes we make (as a temporary hack) so they use
a more default formatting in posts.