Don't process directional keys as grid movement if text selected
In gallery#add_new, if all text in an input box was previously
selected, and the left directional key was pressed, it would move box;
this change means that, if some text is selected (the start and end
selection carets don't match), then the event will be processed
normally instead (and so the directional triggers only occur if there's
no selection in the input box).

Fixes #445.