Add background-color back to login warning
When we removed the inbox flash, we also incidentally
removed the background from the login warning, where you
log in/out in another tab!

This adds it back, piggybacking off the breadcrumbs color.
We could instead use `success`, or `error`, or create a new
separate category for this - I went with a small change
here that hopefully wouldn't use a confusing color, but am
open to thoughts!