Fix alias display when reloading editor
After the NPC box change, the UI will show the base character name upon
loading, even if an alias is selected (and persisting). Selecting an
alias in the UI works fine - it's just when previewing it afterwards
that things show up wrong in the editor - and persisting the alias is
fine too.