Set auth token more correctly, blank session on invalid API requests
Previously the set_auth_token method was called every time we referenced
current_user

Additionally, session[:api_token] is accessed by set_login_gon and
similar, but using string keys - if the user didn't have an auth token
before, and we generate it on this request, the gon value would be nil
incorrectly.

Now that we blank the session on invalid API requests, users can also
reload the page to get a new API token when an invalid request is sent.
This should help get out of bad states more easily!