Avoid clear_section_ids in board section dependent destroy callback
Because of the new transaction behavior, this would now trigger the
UpdateModelJob (in tests) after the board is fully destroyed, but
reliably call the section_id job before the board_id job, triggering an
error when the post tries to handle its callbacks with the board_id
still set but no board present. This behavior may have been present on
old versions and just masked by the test job order!